hasta pto e validado
